+++ AeSys/EoDlgBlockInsert.cpp
@@ -71,7 +71,7 @@ BOOL EoDlgBlockInsert::OnInitDialog() {
 		BlockPosition = m_Document->GetFirstBlockPosition();
 		m_Document->GetNextBlock(BlockPosition, BlockName, Block);
 		SetDlgItemInt(IDC_GROUPS, gsl::narrow_cast<unsigned>(Block->GetCount()), FALSE);
-		SetDlgItemInt(IDC_REFERENCES, m_Document->GetBlockReferenceCount(BlockName), FALSE);
+		SetDlgItemInt(IDC_REFERENCES, static_cast<unsigned>(m_Document->GetBlockReferenceCount(BlockName)), FALSE);
 		WndProcPreviewUpdate(GetDlgItem(IDC_LAYER_PREVIEW)->GetSafeHwnd(), Block);
 	}
 	return TRUE;
@@ -105,7 +105,7 @@ void EoDlgBlockInsert::OnLbnSelchangeBlocksList() {
 		EoDbBlock* Block;
 		m_Document->LookupBlock(BlockName, Block);
 		SetDlgItemInt(IDC_GROUPS, gsl::narrow_cast<unsigned>(Block->GetCount()), FALSE);
-		SetDlgItemInt(IDC_REFERENCES, m_Document->GetBlockReferenceCount(BlockName), FALSE);
+		SetDlgItemInt(IDC_REFERENCES, static_cast<unsigned>(m_Document->GetBlockReferenceCount(BlockName)), FALSE);
 		WndProcPreviewUpdate(GetDlgItem(IDC_LAYER_PREVIEW)->GetSafeHwnd(), Block);
 	}
 }
